move findOrInstallBrowser to utils
